Drivers: fix 32->64 bit pointer/integer conversion warnings
  - acq200.c i2o_handleReadMessageFile()/i2o_handleReadMessage():
    clidata is a void* dual-purposed to carry a 32 bit sample count with
    0xffffffff as the "no data" sentinel.  nsamples is an int, so on
    x86_64 (void*)nsamples sign-extended the sentinel to
    0xffffffffffffffff while the readers in acq200_bridge_read_generic()
    and acq200_fetchDataToLocalBuffer() compare against
    (void*)0xffffffff, i.e. 0x00000000ffffffff.  The comparison could
    never match, so instead of returning -EFAULT the read path fell
    through to count = (size_t)clidata.

    The effect was narrower than the size of that value suggests:
    acq200_copy_from_io_to_user() takes an "int count", so
    0xffffffffffffffff narrowed to -1, its "while (count > 0)" loop never
    ran, and nothing was copied to or from userspace.  The caller then
    did *posp += count - backing the file position up by one - and
    returned count as ssize_t, i.e. -1.  So userspace already saw read()
    fail; it just got a stale errno instead of EFAULT, plus a corrupted
    file position.  Go via u32 so the sentinel round-trips and -EFAULT is
    returned properly.

  - acq32_specific.c: getTargetP() hands back an ioremap'd VA in an
    unsigned long; readw/writew/readl/writel want void __iomem*.  Cast at
    the call sites - the VA is a full 64 bit address, nothing is lost.

  - acq32_core.c: mfa is a 32 bit device-side message frame address kept
    in a void* purely for the 0x%p debug report - widen explicitly.

  - acq32_command.c: do_bucket_command() smuggles argc through a char*
    for help_printf()'s cmd == 0 return path - widen explicitly.

+++ project/BASE/DTACQ/SRC/dt100/Linux/Drivers/acq32_command.c
@@ -1379,7 +1379,7 @@ static int do_bucket_command(
     const char* verb = argv[iarg];
 
     PDEBUGL(1)( " %s\n", argv[0] );
-    HELP_PRINTF_RET( (char*)0, (char*)argc );
+    HELP_PRINTF_RET( (char*)0, (char*)(unsigned long)argc );
     
     strcpy( reject_line, "ERROR in command:" );
     while( iarg != argc ){
